#24 Ability to use the control without the Window
