+++ src/Orc.FilterBuilder/Orc.FilterBuilder.Shared/Views/EditFilterWindow.xaml.cs
@@ -7,17 +7,7 @@
 
 namespace Orc.FilterBuilder.Views
 {
-    using System;
-    using System.Windows.Controls;
-    using System.Windows.Data;
-    using System.Windows.Threading;
-    using Catel.Data;
-    using Catel.IoC;
-    using Catel.Reflection;
     using Catel.Windows;
-    using Converters;
-    using Services;
-    using ViewModels;
 
     /// <summary>
     /// Interaction logic for EditFilterWindow.xaml
@@ -28,47 +18,6 @@ namespace Orc.FilterBuilder.Views
             : base(DataWindowMode.OkCancel, infoBarMessageControlGenerationMode: InfoBarMessageControlGenerationMode.None)
         {
             InitializeComponent();
-        }
-
-        protected override async void OnViewModelChanged()
-        {
-            base.OnViewModelChanged();
-
-            dataGrid.Columns.Clear();
-
-            var vm = ViewModel as EditFilterViewModel;
-            if (vm != null)
-            {
-                if (vm.AllowLivePreview)
-                {
-                    var dependencyResolver = this.GetDependencyResolver();
-                    var reflectionService = dependencyResolver.Resolve<IReflectionService>(vm.FilterScheme.Scope);
-
-                    var targetType = CollectionHelper.GetTargetType(vm.RawCollection);
-                    var instanceProperties = await reflectionService.GetInstancePropertiesAsync(targetType);
-
-                    foreach (var instanceProperty in instanceProperties.Properties)
-                    {
-                        var column = new DataGridTextColumn
-                        {
-                            Header = instanceProperty.DisplayName
-                        };
-
-                        var binding = new Binding
-                        {
-                            Converter = new ObjectToValueConverter(instanceProperty),
-                            ConverterParameter = instanceProperty.Name
-                        };
-
-                        column.Binding = binding;
-
-                        dataGrid.Columns.Add(column);
-                    }
-                }
-
-                // Fix for SA-144
-                var dispatcherOperation = Dispatcher.BeginInvoke(new Action(() => Focus()));
-            }
         }
     }
 }
